ci: use release please for release process
Ports the release process to use release-please as opposed to manual
release. The previous manual release workflow was specific to
the docker image build and release. We will also gain flexibility with
release please as in other Momento SDKs.

In a future PR we will build the binaries and publish to GitHub.